move recommendations over to model
